+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - snapshot_dfn_modify() and properties to responses.SnapshotDefinition
 - Add tcp_ports to DrbdResource
 - snapshot_rollback(..., zfs_rollback_strategy=None)
+- resource_dfn_clone() can now set and delete properties
 
 ### Fixed